organize tools a bit
