+++ src/mcp/tools/core/get_project.ts
