+++ src/mcp/tools/core/get_sdk_config.ts
