+++ src/mcp/tools/core/list_apps.ts
