+++ src/mcp/tools/directory/index.ts
@@ -1,0 +1,4 @@
+import { get_project_directory } from "./get_project_directory";
+import { set_project_directory } from "./set_project_directory";
+
+export const directoryTools = [get_project_directory, set_project_directory];
