+++ src/mcp/types.ts
@@ -1,2 +1,2 @@
-export const SERVER_FEATURES = ["project", "firestore", "storage", "dataconnect", "auth"] as const;
+export const SERVER_FEATURES = ["firestore", "storage", "dataconnect", "auth"] as const;
 export type ServerFeature = (typeof SERVER_FEATURES)[number];
